+++ example/lib/main-localized.dart
@@ -35,6 +35,7 @@ class Demo extends StatelessWidget {
         const Locale('fr', ''), // French, no country code
         const Locale('de', ''), // German, no country code
         const Locale('el', ''), // Greek, no country code
+        const Locale('ht', ''), // Haitian Creole, no country code
         const Locale('hu', ''), // Hungarian, no country code
         const Locale('id', ''), // Indonesian, no country code
         const Locale('it', ''), // Italian, no country code
@@ -128,6 +129,7 @@ class DemoLocalizationsDelegate
         'fr',
         'de',
         'el',
+        'ht',
         'hu',
         'id',
         'it',